+++ ui/src/app/components/pmo-integration-modal/pmo-integration-modal.component.ts
@@ -58,14 +58,14 @@ export class PmoIntegrationModalComponent implements OnInit {
     isConnected: false,
   });
 
-  featuresWithChildren = signal<Ticket[]>([]);
-  expandedFeatureIds = signal<Set<number>>(new Set());
-  expandedPlatformFeatureIds = signal<Set<number>>(new Set());
+  prdsWithChildren = signal<Ticket[]>([]);
+  expandedPrdIds = signal<Set<number>>(new Set());
+  expandedUserStoryIds = signal<Set<number>>(new Set());
 
   // Selected items
-  selectedFeatureIds = signal<Set<number>>(new Set());
-  selectedPlatformFeatureIds = signal<Set<number>>(new Set());
+  selectedPrdIds = signal<Set<number>>(new Set());
   selectedUserStoryIds = signal<Set<number>>(new Set());
+  selectedTaskIds = signal<Set<number>>(new Set());
 
   // Select all state
   allItemsSelected = signal<boolean>(false);
@@ -123,7 +123,7 @@ export class PmoIntegrationModalComponent implements OnInit {
 
       if (validationResult.isValid) {
         this.isLoading.set(false);
-        await this.loadFeaturesHierarchy();
+        await this.loadPrdsHierarchy();
       }
     } catch (error) {
       this.connectionStatus.set({
@@ -138,14 +138,14 @@ export class PmoIntegrationModalComponent implements OnInit {
     }
   }
 
-  async loadFeaturesHierarchy(): Promise<void> {
+  async loadPrdsHierarchy(): Promise<void> {
     try {
       this.isLoadingWorkItems.set(true);
-      const featuresHierarchy = await this.pmoService.getFeaturesHierarchy();
-      this.featuresWithChildren.set(featuresHierarchy);
+      const prdsHierarchy = await this.pmoService.getWorkPlanItemsHierarchy();
+      this.prdsWithChildren.set(prdsHierarchy);
 
-      // Expand all features and platform features by default
-      this.expandAllItemsByDefault(featuresHierarchy);
+      // Expand all PRDs and user stories by default
+      this.expandAllItemsByDefault(prdsHierarchy);
     } catch (error) {
       console.error(
         `Error loading ${this.data.pmoType.toUpperCase()} work items:`,
@@ -159,60 +159,60 @@ export class PmoIntegrationModalComponent implements OnInit {
     }
   }
 
-  private expandAllItemsByDefault(featuresHierarchy: Ticket[]): void {
+  private expandAllItemsByDefault(prdsHierarchy: Ticket[]): void {
     // Create new Sets to store the expanded IDs
-    const featureIds = new Set<number>();
-    const platformFeatureIds = new Set<number>();
+    const prdIds = new Set<number>();
+    const userStoryIds = new Set<number>();
 
-    // Add all feature IDs to the expanded set
-    featuresHierarchy.forEach((feature) => {
-      featureIds.add(parseInt(feature.pmoId));
+    // Add all PRD IDs to the expanded set
+    prdsHierarchy.forEach((prd) => {
+      prdIds.add(parseInt(prd.pmoId));
 
-      // Add all platform feature IDs under each feature to the expanded set
-      if (feature.child && feature.child.length > 0) {
-        feature.child.forEach((platformFeature: Ticket) => {
-          platformFeatureIds.add(parseInt(platformFeature.pmoId));
+      // Add all user story IDs under each PRD to the expanded set
+      if (prd.child && prd.child.length > 0) {
+        prd.child.forEach((userStory: Ticket) => {
+          userStoryIds.add(parseInt(userStory.pmoId));
         });
       }
     });
 
     // Update the signals with all items expanded
-    this.expandedFeatureIds.set(featureIds);
-    this.expandedPlatformFeatureIds.set(platformFeatureIds);
+    this.expandedPrdIds.set(prdIds);
+    this.expandedUserStoryIds.set(userStoryIds);
   }
 
-  toggleFeatureExpansion(featureId: number): void {
-    const currentExpanded = this.expandedFeatureIds();
+  togglePrdExpansion(prdId: number): void {
+    const currentExpanded = this.expandedPrdIds();
     const newExpanded = new Set(currentExpanded);
 
-    if (newExpanded.has(featureId)) {
-      newExpanded.delete(featureId);
+    if (newExpanded.has(prdId)) {
+      newExpanded.delete(prdId);
     } else {
-      newExpanded.add(featureId);
+      newExpanded.add(prdId);
     }
 
-    this.expandedFeatureIds.set(newExpanded);
+    this.expandedPrdIds.set(newExpanded);
   }
 
-  togglePlatformFeatureExpansion(platformFeatureId: number): void {
-    const currentExpanded = this.expandedPlatformFeatureIds();
+  toggleUserStoryExpansion(userStoryId: number): void {
+    const currentExpanded = this.expandedUserStoryIds();
     const newExpanded = new Set(currentExpanded);
 
-    if (newExpanded.has(platformFeatureId)) {
-      newExpanded.delete(platformFeatureId);
+    if (newExpanded.has(userStoryId)) {
+      newExpanded.delete(userStoryId);
     } else {
-      newExpanded.add(platformFeatureId);
+      newExpanded.add(userStoryId);
     }
 
-    this.expandedPlatformFeatureIds.set(newExpanded);
+    this.expandedUserStoryIds.set(newExpanded);
   }
 
-  isFeatureExpanded(featureId: number): boolean {
-    return this.expandedFeatureIds().has(featureId);
+  isPrdExpanded(prdId: number): boolean {
+    return this.expandedPrdIds().has(prdId);
   }
 
-  isPlatformFeatureExpanded(platformFeatureId: number): boolean {
-    return this.expandedPlatformFeatureIds().has(platformFeatureId);
+  isUserStoryExpanded(userStoryId: number): boolean {
+    return this.expandedUserStoryIds().has(userStoryId);
   }
 
   onClose(): void {
@@ -259,81 +259,78 @@ export class PmoIntegrationModalComponent implements OnInit {
     userStories: Ticket[];
     tasks: Ticket[];
   } {
-    const features = this.featuresWithChildren();
+    const prds = this.prdsWithChildren();
 
     // Containers for selected items
-    const prds: Ticket[] = [];
-    const userStories: Ticket[] = [];
-    const tasks: Ticket[] = [];
+    const prdTickets: Ticket[] = [];
+    const userStoryTickets: Ticket[] = [];
+    const taskTickets: Ticket[] = [];
 
-    // Loop through all features (PRDs)
-    features.forEach((feature) => {
-      const featureId = parseInt(feature.pmoId);
+    // Loop through all PRDs
+    prds.forEach((prd: Ticket) => {
+      const prdId = parseInt(prd.pmoId);
 
-      // If this feature is selected, add it to PRDs
-      if (this.isFeatureSelected(featureId)) {
-        // Create a copy of the feature with mapped properties for PRD
-        const prd: Ticket = {
-          pmoId: feature.pmoId,
-          pmoIssueType: feature.pmoIssueType,
-          pmoParentId: feature.pmoParentId,
-          specifaiId: feature.specifaiId || `PRD${feature.pmoId}`,
+      // If this PRD is selected, add it to PRDs
+      if (this.isPrdSelected(prdId)) {
+        // Create a copy of the PRD with mapped properties
+        const prdTicket: Ticket = {
+          pmoId: prd.pmoId,
+          pmoIssueType: prd.pmoIssueType,
+          pmoParentId: prd.pmoParentId,
+          specifaiId: prd.specifaiId || `PRD${prd.pmoId}`,
           specifaiType: 'PRD', // Explicitly set to PRD
           specifaiParentId: null,
-          title: feature.title,
-          description: feature.description,
+          title: prd.title,
+          description: prd.description,
           child: [], // We'll handle the relationships separately
         };
 
-        prds.push(prd);
+        prdTickets.push(prdTicket);
 
-        // Process Platform Features (User Stories) if they exist
-        if (feature.child && feature.child.length > 0) {
-          feature.child.forEach((platformFeature) => {
-            const pfId = parseInt(platformFeature.pmoId);
+        // Process User Stories if they exist
+        if (prd.child && prd.child.length > 0) {
+          prd.child.forEach((userStory: Ticket) => {
+            const usId = parseInt(userStory.pmoId);
 
-            // If this platform feature is selected, add it to User Stories
-            if (this.isPlatformFeatureSelected(pfId)) {
-              // Create a copy of the platform feature with mapped properties for User Story
-              const userStory: Ticket = {
-                pmoId: platformFeature.pmoId,
-                pmoIssueType: platformFeature.pmoIssueType,
-                pmoParentId: platformFeature.pmoParentId,
-                specifaiId:
-                  platformFeature.specifaiId || `US${platformFeature.pmoId}`,
+            // If this user story is selected, add it to User Stories
+            if (this.isUserStorySelected(usId)) {
+              // Create a copy of the user story with mapped properties
+              const userStoryTicket: Ticket = {
+                pmoId: userStory.pmoId,
+                pmoIssueType: userStory.pmoIssueType,
+                pmoParentId: userStory.pmoParentId,
+                specifaiId: userStory.specifaiId || `US${userStory.pmoId}`,
                 specifaiType: 'User Story', // Map to User Story
-                specifaiParentId: prd.specifaiId, // Link to parent PRD
-                title: platformFeature.title,
-                description: platformFeature.description,
+                specifaiParentId: prdTicket.specifaiId, // Link to parent PRD
+                title: userStory.title,
+                description: userStory.description,
                 child: [], // We'll handle the relationships separately
               };
 
-              userStories.push(userStory);
+              userStoryTickets.push(userStoryTicket);
 
-              // Process User Stories (Tasks) if they exist
-              if (platformFeature.child && platformFeature.child.length > 0) {
-                platformFeature.child.forEach((userStory) => {
-                  const usId = parseInt(userStory.pmoId);
+              // Process Tasks if they exist
+              if (userStory.child && userStory.child.length > 0) {
+                userStory.child.forEach((task: Ticket) => {
+                  const taskId = parseInt(task.pmoId);
 
-                  // If this user story is selected, add it to Tasks
-                  if (this.isUserStorySelected(usId)) {
-                    // Create a copy of the user story with mapped properties for Task
-                    const task: Ticket = {
-                      pmoId: userStory.pmoId,
-                      pmoIssueType: userStory.pmoIssueType,
-                      pmoParentId: userStory.pmoParentId,
-                      specifaiId:
-                        userStory.specifaiId || `TASK${userStory.pmoId}`,
+                  // If this task is selected, add it to Tasks
+                  if (this.isTaskSelected(taskId)) {
+                    // Create a copy of the task with mapped properties
+                    const taskTicket: Ticket = {
+                      pmoId: task.pmoId,
+                      pmoIssueType: task.pmoIssueType,
+                      pmoParentId: task.pmoParentId,
+                      specifaiId: task.specifaiId || `TASK${task.pmoId}`,
                       specifaiType: 'Task', // Map to Task
                       specifaiParentId:
-                        platformFeature.specifaiId ||
-                        `US${platformFeature.pmoId}`, // Link to parent User Story
-                      title: userStory.title,
-                      description: userStory.description,
+                        userStory.specifaiId || `US${userStory.pmoId}`, // Link to parent User Story
+                      title: task.title,
+                      description: task.description,
                       child: [], // No children for tasks
                     };
 
-                    tasks.push(task);
+                    taskTickets.push(taskTicket);
                   }
                 });
               }
@@ -344,9 +341,9 @@ export class PmoIntegrationModalComponent implements OnInit {
     });
 
     return {
-      prds,
-      userStories,
-      tasks,
+      prds: prdTickets,
+      userStories: userStoryTickets,
+      tasks: taskTickets,
     };
   }
 
@@ -365,25 +362,25 @@ export class PmoIntegrationModalComponent implements OnInit {
 
     this.allItemsSelected.set(isChecked);
 
-    const features = this.featuresWithChildren();
+    const prds = this.prdsWithChildren();
 
     // New sets for selected items
-    const featureIds = new Set<number>();
-    const platformFeatureIds = new Set<number>();
+    const prdIds = new Set<number>();
     const userStoryIds = new Set<number>();
+    const taskIds = new Set<number>();
 
     if (isChecked) {
       // Add all items to their respective sets
-      features.forEach((feature) => {
-        featureIds.add(parseInt(feature.pmoId));
+      prds.forEach((prd: Ticket) => {
+        prdIds.add(parseInt(prd.pmoId));
 
-        if (feature.child && feature.child.length > 0) {
-          feature.child.forEach((platformFeature: Ticket) => {
-            platformFeatureIds.add(parseInt(platformFeature.pmoId));
+        if (prd.child && prd.child.length > 0) {
+          prd.child.forEach((userStory: Ticket) => {
+            userStoryIds.add(parseInt(userStory.pmoId));
 
-            if (platformFeature.child && platformFeature.child.length > 0) {
-              platformFeature.child.forEach((userStory: Ticket) => {
-                userStoryIds.add(parseInt(userStory.pmoId));
+            if (userStory.child && userStory.child.length > 0) {
+              userStory.child.forEach((task: Ticket) => {
+                taskIds.add(parseInt(task.pmoId));
               });
             }
           });
@@ -392,183 +389,76 @@ export class PmoIntegrationModalComponent implements OnInit {
     }
 
     // Update all selection sets
-    this.selectedFeatureIds.set(featureIds);
-    this.selectedPlatformFeatureIds.set(platformFeatureIds);
+    this.selectedPrdIds.set(prdIds);
     this.selectedUserStoryIds.set(userStoryIds);
+    this.selectedTaskIds.set(taskIds);
   }
 
-  toggleFeatureSelection(event: Event, featureId: number): void {
+  togglePrdSelection(event: Event, prdId: number): void {
     // Stop the click event from propagating to parent elements
     event.stopPropagation();
 
     const checkbox = event.target as HTMLInputElement;
     const isChecked = checkbox.checked;
 
-    const currentSelected = this.selectedFeatureIds();
+    const currentSelected = this.selectedPrdIds();
     const newSelected = new Set(currentSelected);
 
-    // Toggle the feature selection
+    // Toggle the PRD selection
     if (isChecked) {
-      newSelected.add(featureId);
+      newSelected.add(prdId);
     } else {
-      newSelected.delete(featureId);
+      newSelected.delete(prdId);
     }
 
-    this.selectedFeatureIds.set(newSelected);
+    this.selectedPrdIds.set(newSelected);
 
     // Handle cascading selection to children
-    this.cascadeFeatureSelection(featureId, isChecked);
+    this.cascadePrdSelection(prdId, isChecked);
 
     // Update the "all selected" state
     this.updateAllSelectedState();
   }
 
   /**
-   * Cascade feature selection to all its platform features and user stories
+   * Cascade PRD selection to all its user stories and tasks
    */
-  private cascadeFeatureSelection(
-    featureId: number,
-    isSelected: boolean,
-  ): void {
-    const features = this.featuresWithChildren();
-    const feature = features.find((f) => parseInt(f.pmoId) === featureId);
+  private cascadePrdSelection(prdId: number, isSelected: boolean): void {
+    const prds = this.prdsWithChildren();
+    const prd = prds.find((p: Ticket) => parseInt(p.pmoId) === prdId);
 
-    if (!feature || !feature.child) {
+    if (!prd || !prd.child) {
       return;
     }
 
     // Get current selections to modify
-    const platformFeatureIds = new Set(this.selectedPlatformFeatureIds());
     const userStoryIds = new Set(this.selectedUserStoryIds());
+    const taskIds = new Set(this.selectedTaskIds());
 
     // Loop through all children and apply selection
-    feature.child.forEach((platformFeature: Ticket) => {
-      const pfId = parseInt(platformFeature.pmoId);
+    prd.child.forEach((userStory: Ticket) => {
+      const usId = parseInt(userStory.pmoId);
       if (isSelected) {
-        platformFeatureIds.add(pfId);
+        userStoryIds.add(usId);
       } else {
-        platformFeatureIds.delete(pfId);
+        userStoryIds.delete(usId);
       }
 
-      if (platformFeature.child && platformFeature.child.length > 0) {
-        platformFeature.child.forEach((userStory: Ticket) => {
-          const usId = parseInt(userStory.pmoId);
+      if (userStory.child && userStory.child.length > 0) {
+        userStory.child.forEach((task: Ticket) => {
+          const taskId = parseInt(task.pmoId);
           if (isSelected) {
-            userStoryIds.add(usId);
+            taskIds.add(taskId);
           } else {
-            userStoryIds.delete(usId);
+            taskIds.delete(taskId);
           }
         });
       }
     });
 
     // Update the selection sets
-    this.selectedPlatformFeatureIds.set(platformFeatureIds);
     this.selectedUserStoryIds.set(userStoryIds);
-  }
-
-  togglePlatformFeatureSelection(
-    event: Event,
-    platformFeatureId: number,
-  ): void {
-    // Stop the click event from propagating to parent elements
-    event.stopPropagation();
-
-    const checkbox = event.target as HTMLInputElement;
-    const isChecked = checkbox.checked;
-
-    const currentSelected = this.selectedPlatformFeatureIds();
-    const newSelected = new Set(currentSelected);
-
-    // Toggle the platform feature selection
-    if (isChecked) {
-      newSelected.add(platformFeatureId);
-
-      // If selected, also select parent feature
-      this.selectParentFeature(platformFeatureId);
-    } else {
-      newSelected.delete(platformFeatureId);
-    }
-
-    this.selectedPlatformFeatureIds.set(newSelected);
-
-    // Handle cascading selection to children user stories
-    // Always cascade selection/deselection to children
-    this.cascadePlatformFeatureSelection(platformFeatureId, isChecked);
-
-    // Update the "all selected" state
-    this.updateAllSelectedState();
-  }
-
-  /**
-   * Find and select the parent feature of a platform feature
-   */
-  private selectParentFeature(platformFeatureId: number): void {
-    const features = this.featuresWithChildren();
-    let parentFeature: Ticket | null = null;
-
-    // Find the platform feature's parent feature
-    for (const feature of features) {
-      if (!feature.child) continue;
-
-      const hasPlatformFeature = feature.child.some(
-        (pf: Ticket) => parseInt(pf.pmoId) === platformFeatureId,
-      );
-      if (hasPlatformFeature) {
-        parentFeature = feature;
-        break;
-      }
-    }
-
-    if (parentFeature) {
-      // Select the parent feature
-      const featureIds = new Set(this.selectedFeatureIds());
-      featureIds.add(parseInt(parentFeature.pmoId));
-      this.selectedFeatureIds.set(featureIds);
-    }
-  }
-
-  /**
-   * Cascade platform feature selection to all its user stories
-   */
-  private cascadePlatformFeatureSelection(
-    platformFeatureId: number,
-    isSelected: boolean,
-  ): void {
-    // Find the platform feature in the hierarchy
-    let foundPlatformFeature: Ticket | undefined;
-
-    const features = this.featuresWithChildren();
-    for (const feature of features) {
-      if (!feature.child) continue;
-
-      const pf = feature.child.find(
-        (pf: Ticket) => parseInt(pf.pmoId) === platformFeatureId,
-      );
-      if (pf) {
-        foundPlatformFeature = pf;
-        break;
-      }
-    }
-
-    if (!foundPlatformFeature || !foundPlatformFeature.child) {
-      return;
-    }
-
-    // Get current selections to modify
-    const userStoryIds = new Set(this.selectedUserStoryIds());
-
-    // Apply selection to all user stories
-    foundPlatformFeature.child.forEach((userStory: Ticket) => {
-      if (isSelected) {
-        userStoryIds.add(parseInt(userStory.pmoId));
-      } else {
-        userStoryIds.delete(parseInt(userStory.pmoId));
-      }
-    });
-
-    // Update the selection set
-    this.selectedUserStoryIds.set(userStoryIds);
+    this.selectedTaskIds.set(taskIds);
   }
 
   toggleUserStorySelection(event: Event, userStoryId: number): void {
@@ -585,69 +475,170 @@ export class PmoIntegrationModalComponent implements OnInit {
     if (isChecked) {
       newSelected.add(userStoryId);
 
-      // If selected, select all parents in hierarchy (platform feature and feature)
-      this.selectParentHierarchy(userStoryId);
+      // If selected, also select parent PRD
+      this.selectParentPrd(userStoryId);
     } else {
       newSelected.delete(userStoryId);
     }
 
     this.selectedUserStoryIds.set(newSelected);
 
+    // Handle cascading selection to children tasks
+    // Always cascade selection/deselection to children
+    this.cascadeUserStorySelection(userStoryId, isChecked);
+
     // Update the "all selected" state
     this.updateAllSelectedState();
   }
 
   /**
-   * Automatically selects the entire parent hierarchy of a user story
-   * (both platform feature and feature)
+   * Find and select the parent PRD of a user story
    */
-  private selectParentHierarchy(userStoryId: number): void {
-    const features = this.featuresWithChildren();
-    let foundPlatformFeature: Ticket | undefined = undefined;
-    let foundFeature: Ticket | undefined = undefined;
+  private selectParentPrd(userStoryId: number): void {
+    const prds = this.prdsWithChildren();
+    let parentPrd: Ticket | null = null;
 
-    // Find the user story's parent platform feature and grandparent feature
-    for (const feature of features) {
-      if (!feature.child) continue;
+    // Find the user story's parent PRD
+    for (const prd of prds) {
+      if (!prd.child) continue;
 
-      for (const platformFeature of feature.child) {
-        if (!platformFeature.child) continue;
+      const hasUserStory = prd.child.some(
+        (us: Ticket) => parseInt(us.pmoId) === userStoryId,
+      );
+      if (hasUserStory) {
+        parentPrd = prd;
+        break;
+      }
+    }
 
-        const foundUserStory = platformFeature.child.find(
-          (us: Ticket) => parseInt(us.pmoId) === userStoryId,
+    if (parentPrd) {
+      // Select the parent PRD
+      const prdIds = new Set(this.selectedPrdIds());
+      prdIds.add(parseInt(parentPrd.pmoId));
+      this.selectedPrdIds.set(prdIds);
+    }
+  }
+
+  /**
+   * Cascade user story selection to all its tasks
+   */
+  private cascadeUserStorySelection(
+    userStoryId: number,
+    isSelected: boolean,
+  ): void {
+    // Find the user story in the hierarchy
+    let foundUserStory: Ticket | undefined;
+
+    const prds = this.prdsWithChildren();
+    for (const prd of prds) {
+      if (!prd.child) continue;
+
+      const us = prd.child.find(
+        (us: Ticket) => parseInt(us.pmoId) === userStoryId,
+      );
+      if (us) {
+        foundUserStory = us;
+        break;
+      }
+    }
+
+    if (!foundUserStory || !foundUserStory.child) {
+      return;
+    }
+
+    // Get current selections to modify
+    const taskIds = new Set(this.selectedTaskIds());
+
+    // Apply selection to all tasks
+    foundUserStory.child.forEach((task: Ticket) => {
+      if (isSelected) {
+        taskIds.add(parseInt(task.pmoId));
+      } else {
+        taskIds.delete(parseInt(task.pmoId));
+      }
+    });
+
+    // Update the selection set
+    this.selectedTaskIds.set(taskIds);
+  }
+
+  toggleTaskSelection(event: Event, taskId: number): void {
+    // Stop the click event from propagating to parent elements
+    event.stopPropagation();
+
+    const checkbox = event.target as HTMLInputElement;
+    const isChecked = checkbox.checked;
+
+    const currentSelected = this.selectedTaskIds();
+    const newSelected = new Set(currentSelected);
+
+    // Toggle the task selection
+    if (isChecked) {
+      newSelected.add(taskId);
+
+      // If selected, select all parents in hierarchy (user story and PRD)
+      this.selectParentHierarchyForTask(taskId);
+    } else {
+      newSelected.delete(taskId);
+    }
+
+    this.selectedTaskIds.set(newSelected);
+
+    // Update the "all selected" state
+    this.updateAllSelectedState();
+  }
+
+  /**
+   * Automatically selects the entire parent hierarchy of a task
+   * (both user story and PRD)
+   */
+  private selectParentHierarchyForTask(taskId: number): void {
+    const prds = this.prdsWithChildren();
+    let foundUserStory: Ticket | undefined = undefined;
+    let foundPrd: Ticket | undefined = undefined;
+
+    // Find the task's parent user story and grandparent PRD
+    for (const prd of prds) {
+      if (!prd.child) continue;
+
+      for (const userStory of prd.child) {
+        if (!userStory.child) continue;
+
+        const foundTask = userStory.child.find(
+          (t: Ticket) => parseInt(t.pmoId) === taskId,
         );
-        if (foundUserStory) {
-          foundPlatformFeature = platformFeature;
-          foundFeature = feature;
+        if (foundTask) {
+          foundUserStory = userStory;
+          foundPrd = prd;
           break;
         }
       }
-      if (foundPlatformFeature) break;
+      if (foundUserStory) break;
     }
 
-    if (foundPlatformFeature && foundFeature) {
-      // Select the platform feature
-      const platformFeatureIds = new Set(this.selectedPlatformFeatureIds());
-      platformFeatureIds.add(parseInt(foundPlatformFeature.pmoId));
-      this.selectedPlatformFeatureIds.set(platformFeatureIds);
+    if (foundUserStory && foundPrd) {
+      // Select the user story
+      const userStoryIds = new Set(this.selectedUserStoryIds());
+      userStoryIds.add(parseInt(foundUserStory.pmoId));
+      this.selectedUserStoryIds.set(userStoryIds);
 
-      // Select the feature
-      const featureIds = new Set(this.selectedFeatureIds());
-      featureIds.add(parseInt(foundFeature.pmoId));
-      this.selectedFeatureIds.set(featureIds);
+      // Select the PRD
+      const prdIds = new Set(this.selectedPrdIds());
+      prdIds.add(parseInt(foundPrd.pmoId));
+      this.selectedPrdIds.set(prdIds);
     }
   }
 
-  isFeatureSelected(featureId: number): boolean {
-    return this.selectedFeatureIds().has(featureId);
-  }
-
-  isPlatformFeatureSelected(platformFeatureId: number): boolean {
-    return this.selectedPlatformFeatureIds().has(platformFeatureId);
+  isPrdSelected(prdId: number): boolean {
+    return this.selectedPrdIds().has(prdId);
   }
 
   isUserStorySelected(userStoryId: number): boolean {
     return this.selectedUserStoryIds().has(userStoryId);
+  }
+
+  isTaskSelected(taskId: number): boolean {
+    return this.selectedTaskIds().has(taskId);
   }
 
   // Helper method to make parseInt accessible from the template
@@ -660,24 +651,24 @@ export class PmoIntegrationModalComponent implements OnInit {
     let totalItemCount = 0;
     let selectedItemCount = 0;
 
-    const features = this.featuresWithChildren();
-    features.forEach((feature) => {
-      totalItemCount++; // Count the feature
-      if (this.isFeatureSelected(parseInt(feature.pmoId))) {
+    const prds = this.prdsWithChildren();
+    prds.forEach((prd: Ticket) => {
+      totalItemCount++; // Count the PRD
+      if (this.isPrdSelected(parseInt(prd.pmoId))) {
         selectedItemCount++;
       }
 
-      if (feature.child && feature.child.length > 0) {
-        feature.child.forEach((platformFeature: Ticket) => {
-          totalItemCount++; // Count the platform feature
-          if (this.isPlatformFeatureSelected(parseInt(platformFeature.pmoId))) {
+      if (prd.child && prd.child.length > 0) {
+        prd.child.forEach((userStory: Ticket) => {
+          totalItemCount++; // Count the user story
+          if (this.isUserStorySelected(parseInt(userStory.pmoId))) {
             selectedItemCount++;
           }
 
-          if (platformFeature.child && platformFeature.child.length > 0) {
-            platformFeature.child.forEach((userStory: Ticket) => {
-              totalItemCount++; // Count the user story
-              if (this.isUserStorySelected(parseInt(userStory.pmoId))) {
+          if (userStory.child && userStory.child.length > 0) {
+            userStory.child.forEach((task: Ticket) => {
+              totalItemCount++; // Count the task
+              if (this.isTaskSelected(parseInt(task.pmoId))) {
                 selectedItemCount++;
               }
             });
