+++ ui/src/app/integrations/ado/ado.service.ts
@@ -1,11 +1,15 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { lastValueFrom, take } from 'rxjs';
+import { lastValueFrom, take, first } from 'rxjs';
 import { Ticket } from '../../services/pmo-integration/pmo-integration.service';
 import { PmoService } from '../../services/pmo-integration/pmo-service.interface';
 import { ElectronService } from '../../electron-bridge/electron.service';
 import { Store } from '@ngxs/store';
 import { ProjectsState } from 'src/app/store/projects/projects.state';
+import { BulkUpdateFiles, GetProjectFiles, ReadFile } from '../../store/projects/projects.actions';
+import { ToasterService } from '../../services/toaster/toaster.service';
+import { Router } from '@angular/router';
+import { EXPORT_FILE_FORMATS, ExportFileFormat } from '../../constants/export.constants';
 
 interface AdoConfiguration {
   personalAccessToken: string;
@@ -24,6 +28,8 @@ export class AdoService implements PmoService {
     private http: HttpClient,
     private electronService: ElectronService,
     private store: Store,
+    private toast: ToasterService,
+    private router: Router,
   ) {}
 
   /**
@@ -115,7 +121,7 @@ export class AdoService implements PmoService {
    * This creates a three-level hierarchy that can be displayed in a tree view
    * Mapped to the Ticket interface for PMO integration
    */
-  async getFeaturesHierarchy(): Promise<Ticket[]> {
+  async getWorkPlanItemsHierarchy(): Promise<Ticket[]> {
     if (!this.config || !this.baseUrl) {
       console.error(
         'Azure DevOps service not configured. Call configure() first.',
@@ -403,5 +409,604 @@ export class AdoService implements PmoService {
       console.error('Error executing WIQL query:', error);
       throw error;
     }
+  }
+
+  /**
+   * Execute the actual ADO action (pull or push)
+   */
+  executeAdoAction(folderName: string, action: 'pull' | 'push'): void {
+    // For now, we'll use the existing export functionality
+    // TODO: Implement actual ADO pull/push logic
+    this.exportDocumentList(folderName, EXPORT_FILE_FORMATS.JSON);
+    
+    const actionText = action === 'pull' ? 'Pulling from' : 'Pushing to';
+    this.toast.showSuccess(`${actionText} ADO for ${folderName}`);
+  }
+
+  /**
+   * Export document list (placeholder for actual export functionality)
+   */
+  private exportDocumentList(folder: string, format: ExportFileFormat): void {
+    // This would need to be implemented to actually export the document list
+    // For now, it's a placeholder
+    console.log(`Exporting ${folder} in ${format} format`);
+  }
+
+  /**
+   * Navigate to ADO configuration page or open settings
+   */
+  navigateToAdoConfiguration(appInfo: any): void {
+    // Create the state object to pass with navigation
+    const navState = { 
+      data: appInfo,
+      selectedFolder: {
+        title: 'app-integrations',
+        id: appInfo?.id,
+        metadata: appInfo
+      },
+      selectedIntegration: 'ado',
+      openPmoAccordion: true 
+    };
+    
+    // Check if we're already on the apps/{id} page
+    const currentUrl = this.router.url;
+    const targetUrl = `/apps/${appInfo?.id}`;
+    
+    if (currentUrl === targetUrl) {
+      const event = new CustomEvent('open-pmo-integration', { detail: navState });
+      window.dispatchEvent(event);
+    } else {
+      this.router.navigate([targetUrl], { state: navState });
+    }
+  }
+
+  /**
+   * Process selected ADO items and create/update PRD files
+   * @param folderName The folder name 
+   * @param action The action (pull or push)
+   * @param selectedItems The items selected in the ADO integration modal
+   * @param appInfo The application info/metadata
+   */
+  processAdoSelectedItems(
+    folderName: string, 
+    action: 'pull' | 'push', 
+    selectedItems: { 
+      prds: Ticket[], 
+      userStories: Ticket[], 
+      tasks: Ticket[] 
+    },
+    appInfo: any
+  ): void {
+    console.log(`Processing ADO items for ${folderName} with action: ${action}`, selectedItems);
+    console.log("Project metadata:", appInfo);
+    
+    if (action === 'pull') {
+      try {
+        // First, get the existing PRD files to check for pmoId and get max PRD number
+        this.getExistingPrdFiles(appInfo).then(existingData => {
+          const { maxPrdNumber, existingPmoMap, existingFeatureFiles } = existingData;
+          
+          // Group tickets by parent relationship
+          const ticketGroups = this.groupTicketsByParentRelationship(selectedItems);
+          
+          // Prepare updates array for BulkUpdateFiles action
+          const updates: { path: string; content: any }[] = [];
+          
+          // Process each PRD (Feature from ADO)
+          ticketGroups.forEach((group, index) => {
+            const prd = group.prd;
+            
+            // Check if this PRD already exists based on pmoId
+            const existingPrd = existingPmoMap.prds[prd.pmoId];
+            
+            // Determine PRD ID - use existing or generate new one with zero-padding (PRD01, PRD02, etc.)
+            const prdId = existingPrd ? 
+              existingPrd.specifaiId : 
+              `PRD${String(maxPrdNumber + index + 1).padStart(2, '0')}`;
+            
+            // Format the user stories and tasks for this PRD
+            const features = this.formatFeaturesForPrd(group.userStories, group.tasks, existingPmoMap);
+            
+            // Create the PRD base file content with correct structure
+            const prdBaseContent = {
+              id: prdId,
+              title: prd.title,
+              requirement: prd.description || '', // Map description to requirement field
+              state: 'Active',
+              createdAt: existingPrd ? new Date().toISOString() : new Date().toISOString(),
+              updatedAt: new Date().toISOString(),
+              pmoId: prd.pmoId, // Store ADO ID for future reference
+              pmoIssueType: prd.pmoIssueType,
+              chatHistory: [] // Add empty chatHistory array as required
+            };
+            
+            // Check if there's an existing feature file for this PRD
+            const featureFilePath = `prd/${prdId}-feature.json`;
+            const existingFeatureFile = existingFeatureFiles[prdId];
+            
+            let prdFeatureContent: any = { features: features };
+            
+            // If this PRD exists and we have the feature file content, merge with new features
+            if (existingPrd && existingFeatureFile) {
+              const existingFeatures = existingFeatureFile.features || [];
+              
+              // Create a map of existing features by pmoId for easy lookup
+              const existingFeatureMap = new Map<string, any>();
+              existingFeatures.forEach((feature: any) => {
+                if (feature.pmoId) {
+                  existingFeatureMap.set(feature.pmoId, feature);
+                }
+              });
+              
+              // Merge or update features
+              const mergedFeatures = features.map(newFeature => {
+                const existingFeature = existingFeatureMap.get(newFeature.pmoId);
+                
+                // If this feature already exists, update its properties but preserve any
+                // that might be set in the UI and not coming from ADO
+                if (existingFeature) {
+                  existingFeatureMap.delete(newFeature.pmoId); // Remove from map to track what's processed
+                  
+                  // Create a map of existing tasks by pmoId
+                  const existingTaskMap = new Map<string, any>();
+                  (existingFeature.tasks || []).forEach((task: any) => {
+                    if (task.pmoId) {
+                      existingTaskMap.set(task.pmoId, task);
+                    }
+                  });
+                  
+                  // Merge tasks within this feature
+                  const mergedTasks = (newFeature.tasks || []).map((newTask: any) => {
+                    const existingTask = existingTaskMap.get(newTask.pmoId);
+                    
+                    // If task exists, update it
+                    if (existingTask) {
+                      existingTaskMap.delete(newTask.pmoId); // Remove from map to track what's processed
+                      return {
+                        ...existingTask,
+                        list: newTask.list,
+                        acceptance: newTask.acceptance || existingTask.acceptance,
+                        // Keep other properties from existing task
+                      };
+                    }
+                    
+                    // This is a new task
+                    return newTask;
+                  });
+                  
+                  // Add any remaining existing tasks that weren't in the new data
+                  existingTaskMap.forEach(remainingTask => {
+                    mergedTasks.push(remainingTask);
+                  });
+                  
+                  // Return the merged feature
+                  return {
+                    ...existingFeature,
+                    name: newFeature.name,
+                    description: newFeature.description || existingFeature.description,
+                    tasks: mergedTasks,
+                    // Keep other properties from existing feature
+                  };
+                }
+                
+                // This is a new feature
+                return newFeature;
+              });
+              
+              // Add any remaining existing features that weren't in the new data
+              existingFeatureMap.forEach(remainingFeature => {
+                mergedFeatures.push(remainingFeature);
+              });
+              
+              // Update the feature content with merged features
+              prdFeatureContent = { 
+                ...existingFeatureFile, 
+                features: mergedFeatures 
+              };
+            }
+            
+            // Log what we're saving for debugging
+            console.log(`Creating/updating PRD base file: ${prdId}-base.json`, prdBaseContent);
+            console.log(`Creating/updating PRD feature file: ${prdId}-feature.json`, prdFeatureContent);
+            
+            // Add PRD files to updates
+            updates.push({
+              path: `prd/${prdId}-base.json`,
+              content: prdBaseContent
+            });
+            
+            updates.push({
+              path: featureFilePath,
+              content: prdFeatureContent
+            });
+          });
+          
+          // If we have updates, dispatch the BulkUpdateFiles action
+          if (updates.length > 0) {
+            // Calculate the new max PRD number after adding all new PRDs
+            const newMaxPrdNumber = maxPrdNumber + ticketGroups.filter(group => {
+              const prd = group.prd;
+              return !existingPmoMap.prds[prd.pmoId]; // Only count new PRDs
+            }).length;
+            
+            // Update the metadata with the new PRD count
+            this.updatePrdCountInMetadata(newMaxPrdNumber, appInfo);
+            
+            this.store.dispatch(new BulkUpdateFiles(updates));
+            this.toast.showSuccess(`Successfully imported ${ticketGroups.length} PRDs with their user stories and tasks from ADO`);
+            
+            // Refresh the file list
+            const projectId = appInfo?.id;
+            if (projectId) {
+              this.store.dispatch(new GetProjectFiles(projectId));
+            }
+          } else {
+            this.toast.showInfo('No items to import from ADO');
+          }
+        }).catch(error => {
+          console.error('Error processing ADO items:', error);
+          this.toast.showError('Failed to process ADO items');
+        });
+      } catch (error) {
+        console.error('Error importing items from ADO:', error);
+        this.toast.showError('Failed to import items from ADO');
+      }
+    } else {
+      // For 'push' action (not implemented yet)
+      this.toast.showInfo('Push to ADO functionality is not yet implemented');
+    }
+  }
+
+  /**
+   * Fetch existing PRD files and analyze them to find max PRD number and existing pmoIds
+   * @returns Promise with information about existing PRD files
+   */
+  private getExistingPrdFiles(appInfo: any): Promise<{
+    maxPrdNumber: number;
+    existingPmoMap: {
+      prds: { [pmoId: string]: { specifaiId: string; path: string } };
+      userStories: { [pmoId: string]: { specifaiId: string; path: string } };
+      tasks: { [pmoId: string]: { specifaiId: string; path: string } };
+    };
+    existingFeatureFiles: { [prdId: string]: any };
+  }> {
+    return new Promise((resolve, reject) => {
+      // Get all PRD files from the project
+      this.store.select(ProjectsState.getProjectsFolders).pipe(first()).subscribe(
+        directories => {
+          try {
+            // Find the PRD directory
+            const prdDir = directories.find(dir => dir.name === 'prd');
+            
+            if (!prdDir || !prdDir.children || prdDir.children.length === 0) {
+              // No PRD files exist yet
+              resolve({
+                maxPrdNumber: 0,
+                existingPmoMap: { prds: {}, userStories: {}, tasks: {} },
+                existingFeatureFiles: {}
+              });
+              return;
+            }
+            
+            // Get all feature files (PRD*-feature.json)
+            const featureFiles = prdDir.children.filter(file => 
+              file.includes('-feature.json') && !file.includes('-archived')
+            );
+            
+            // Get all base files (PRD*-base.json)
+            const baseFiles = prdDir.children.filter(file => 
+              file.includes('-base.json') && !file.includes('-archived')
+            );
+            
+            // Find max PRD number - use the regex to match PRD with any number format
+            let maxPrdNumber = 0;
+            const prdRegex = /PRD(\d+)[-]/;
+
+            console.log("Project metadata for PRD count:", appInfo);
+            
+            // Check if we have project metadata with prdCount
+            if (appInfo && appInfo.prdCount) {
+              // Use the value from metadata.json
+              maxPrdNumber = parseInt(appInfo.prdCount, 10);
+            } else {
+              // Fall back to scanning file names
+              baseFiles.forEach(file => {
+                const match = file.match(prdRegex);
+                if (match && match[1]) {
+                  // Handle both PRD1 and PRD01 formats by stripping leading zeros
+                  const prdNumber = parseInt(match[1], 10);
+                  if (prdNumber > maxPrdNumber) {
+                    maxPrdNumber = prdNumber;
+                  }
+                }
+              });
+            }
+            
+            // Initialize mapping objects for existing items with pmoId
+            const existingPmoMap = {
+              prds: {} as { [pmoId: string]: { specifaiId: string; path: string } },
+              userStories: {} as { [pmoId: string]: { specifaiId: string; path: string } },
+              tasks: {} as { [pmoId: string]: { specifaiId: string; path: string } }
+            };
+            
+            // Object to store existing feature files content
+            const existingFeatureFiles: { [prdId: string]: any } = {};
+            
+            // Process base files to find PRDs with pmoId
+            const baseFilePromises = baseFiles.map(file => {
+              return new Promise<void>((resolveFile) => {
+                // Extract PRD ID from filename
+                const prdMatch = file.match(prdRegex);
+                if (!prdMatch || !prdMatch[1]) {
+                  resolveFile();
+                  return;
+                }
+                
+                // Ensure consistent format with zero-padding (PRD01, PRD02, etc.)
+                const prdNumber = parseInt(prdMatch[1], 10);
+                const prdId = `PRD${String(prdNumber).padStart(2, '0')}`;
+                
+                // Read the base file to check for pmoId
+                this.store.dispatch(new ReadFile(`prd/${file}`))
+                  .pipe(first())
+                  .subscribe({
+                    next: () => {
+                      // Get the file content from state
+                      this.store.select(ProjectsState.getSelectedFileContent)
+                        .pipe(first())
+                        .subscribe(content => {
+                          // Check for pmoId in both formats (old and new structure)
+                          if (content && content.pmoId) {
+                            // Store mapping of pmoId to specifaiId for PRDs
+                            existingPmoMap.prds[content.pmoId] = {
+                              specifaiId: prdId,
+                              path: `prd/${file}`
+                            };
+                          }
+                          resolveFile();
+                        });
+                    },
+                    error: () => {
+                      // Ignore errors and continue
+                      resolveFile();
+                    }
+                  });
+              });
+            });
+            
+            // Process feature files to find user stories and tasks with pmoId
+            const featureFilePromises = featureFiles.map(file => {
+              return new Promise<void>((resolveFile) => {
+                // Extract PRD ID from filename
+                const prdMatch = file.match(prdRegex);
+                if (!prdMatch || !prdMatch[1]) {
+                  resolveFile();
+                  return;
+                }
+                
+                // Ensure consistent format with zero-padding (PRD01, PRD02, etc.)
+                const prdNumber = parseInt(prdMatch[1], 10);
+                const prdId = `PRD${String(prdNumber).padStart(2, '0')}`;
+                
+                // Read the feature file to check for user stories and tasks with pmoId
+                this.store.dispatch(new ReadFile(`prd/${file}`))
+                  .pipe(first())
+                  .subscribe({
+                    next: () => {
+                      // Get the file content from state
+                      this.store.select(ProjectsState.getSelectedFileContent)
+                        .pipe(first())
+                        .subscribe(content => {
+                          // Store the feature file content
+                          existingFeatureFiles[prdId] = content;
+                          
+                          // Process features (user stories) and their tasks
+                          if (content && content.features && Array.isArray(content.features)) {
+                            content.features.forEach((feature: any) => {
+                              if (feature.pmoId) {
+                                // Store mapping of pmoId to specifaiId for user stories
+                                existingPmoMap.userStories[feature.pmoId] = {
+                                  specifaiId: feature.id,
+                                  path: `prd/${file}`
+                                };
+                              }
+                              
+                              // Process tasks
+                              if (feature.tasks && Array.isArray(feature.tasks)) {
+                                feature.tasks.forEach((task: any) => {
+                                  if (task.pmoId) {
+                                    // Store mapping of pmoId to specifaiId for tasks
+                                    existingPmoMap.tasks[task.pmoId] = {
+                                      specifaiId: task.id,
+                                      path: `prd/${file}`
+                                    };
+                                  }
+                                });
+                              }
+                            });
+                          }
+                          resolveFile();
+                        });
+                    },
+                    error: () => {
+                      // Ignore errors and continue
+                      resolveFile();
+                    }
+                  });
+              });
+            });
+            
+            // Wait for all file reading to complete
+            Promise.all([...baseFilePromises, ...featureFilePromises]).then(() => {
+              resolve({
+                maxPrdNumber,
+                existingPmoMap,
+                existingFeatureFiles
+              });
+            }).catch(error => {
+              reject(error);
+            });
+          } catch (error) {
+            reject(error);
+          }
+        },
+        error => reject(error)
+      );
+    });
+  }
+  
+  /**
+   * Update the PRD count in the project metadata if needed
+   * @param newPrdCount The new PRD count
+   * @param appInfo The application info/metadata
+   */
+  private updatePrdCountInMetadata(newPrdCount: number, appInfo: any): void {
+    // Only update if we have appInfo and the new count is higher than the current one
+    if (appInfo) {
+      const currentCount = appInfo.prdCount ? parseInt(appInfo.prdCount, 10) : 0;
+      
+      if (newPrdCount > currentCount) {
+        // Create a copy of the metadata to update
+        const updatedMetadata = { 
+          ...appInfo, 
+          prdCount: String(newPrdCount) 
+        };
+        
+        // We should update the metadata in the project, but this would require
+        // invoking the electron service to update the .metadata.json file
+        // For now, just update our local copy
+        // TODO: Implement updating metadata file
+      }
+    }
+  }
+  
+  /**
+   * Group tickets by their parent-child relationships
+   * @param selectedItems Selected tickets from ADO
+   * @returns Array of ticket groups, each containing a PRD and its related user stories and tasks
+   */
+  private groupTicketsByParentRelationship(selectedItems: {
+    prds: Ticket[];
+    userStories: Ticket[];
+    tasks: Ticket[];
+  }): Array<{
+    prd: Ticket;
+    userStories: Ticket[];
+    tasks: { [userStoryId: string]: Ticket[] };
+  }> {
+    const ticketGroups: Array<{
+      prd: Ticket;
+      userStories: Ticket[];
+      tasks: { [userStoryId: string]: Ticket[] };
+    }> = [];
+    
+    // Process each PRD
+    selectedItems.prds.forEach(prd => {
+      // Find user stories for this PRD
+      const relatedUserStories = selectedItems.userStories.filter(
+        us => us.pmoParentId === prd.pmoId
+      );
+      
+      // Create task map for this PRD
+      const taskMap: { [userStoryId: string]: Ticket[] } = {};
+      
+      // Process each user story to find related tasks
+      relatedUserStories.forEach(userStory => {
+        // Find tasks for this user story
+        const relatedTasks = selectedItems.tasks.filter(
+          task => task.pmoParentId === userStory.pmoId
+        );
+        
+        if (relatedTasks.length > 0) {
+          taskMap[userStory.pmoId] = relatedTasks;
+        }
+      });
+      
+      // Add this PRD group
+      ticketGroups.push({
+        prd,
+        userStories: relatedUserStories,
+        tasks: taskMap
+      });
+    });
+    
+    return ticketGroups;
+  }
+  
+  /**
+   * Format user stories and tasks into the features format required for PRD feature files
+   * @param userStories User stories to include in the PRD
+   * @param tasks Tasks mapped by user story ID
+   * @param existingPmoMap Map of existing items with pmoId
+   * @returns Array of formatted features
+   */
+  private formatFeaturesForPrd(
+    userStories: Ticket[], 
+    tasks: { [userStoryId: string]: Ticket[] },
+    existingPmoMap: {
+      prds: { [pmoId: string]: { specifaiId: string; path: string } };
+      userStories: { [pmoId: string]: { specifaiId: string; path: string } };
+      tasks: { [pmoId: string]: { specifaiId: string; path: string } };
+    }
+  ): Array<any> {
+    let nextUserStoryId = 1;
+    let nextTaskId = 1;
+    
+    // Find the highest existing US and TASK IDs
+    Object.values(existingPmoMap.userStories).forEach(item => {
+      const match = item.specifaiId.match(/US(\d+)/);
+      if (match && match[1]) {
+        const id = parseInt(match[1], 10);
+        if (id >= nextUserStoryId) {
+          nextUserStoryId = id + 1;
+        }
+      }
+    });
+    
+    Object.values(existingPmoMap.tasks).forEach(item => {
+      const match = item.specifaiId.match(/TASK(\d+)/);
+      if (match && match[1]) {
+        const id = parseInt(match[1], 10);
+        if (id >= nextTaskId) {
+          nextTaskId = id + 1;
+        }
+      }
+    });
+    
+    // Format user stories into features
+    return userStories.map(userStory => {
+      // Determine user story ID - use existing or generate new one
+      const userStoryId = existingPmoMap.userStories[userStory.pmoId] ?
+        existingPmoMap.userStories[userStory.pmoId].specifaiId :
+        `US${nextUserStoryId++}`;
+      
+      // Get tasks for this user story
+      const userStoryTasks = tasks[userStory.pmoId] || [];
+      
+      // Format tasks
+      const formattedTasks = userStoryTasks.map(task => {
+        // Determine task ID - use existing or generate new one
+        const taskId = existingPmoMap.tasks[task.pmoId] ?
+          existingPmoMap.tasks[task.pmoId].specifaiId :
+          `TASK${nextTaskId++}`;
+        
+        return {
+          id: taskId,
+          list: task.title,
+          acceptance: task.description || '',
+          status: 'Active',
+          pmoId: task.pmoId // Store ADO ID for future reference
+        };
+      });
+      
+      // Return the feature object (user story)
+      return {
+        id: userStoryId,
+        name: userStory.title,
+        description: userStory.description || '',
+        pmoId: userStory.pmoId, // Store ADO ID for future reference
+        tasks: formattedTasks
+      };
+    });
   }
 }
